Crate d3d11

Rust bindings for D3D11

Reexports

pub use core::enumerations::*;

pub use core::structures::*;

pub use core::functions::*;

pub use core::interfaces::*;

pub use common_version::enumerations::*;

pub use common_version::structures::*;

pub use common_version::interfaces::*;

pub use resource::enumerations::*;

pub use resource::structures::*;

pub use resource::interfaces::*;

pub use resource::functions::*;

pub use shader::enumerations::*;

pub use shader::structures::*;

pub use shader::interfaces::*;

pub use constants::*;

Modules

common_version
"The Direct3D API defines several API elements that are common to the Direct3D 11, Direct3D 10, and Direct3D 10.1 versions, and versions later than Direct3D 11. You can use these API elements in your code for any of these Direct3D versions. These API elements are known as version neutral."

constants
Constants provided by D3D11

core
Core components

interfaces

resource
"The Direct3D API defines several API elements to help you create and manage resources"

shader
Several API elements to help create and manage programmable shaders